+++ packages/audioplayers_darwin/darwin/audioplayers_darwin/Sources/audioplayers_darwin/AudioPlayerError.swift
@@ -1,4 +1,4 @@
-public enum AudioPlayerError: Error {
+enum AudioPlayerError: Error {
   case error(String)
   case warning(String)
 }
