+++ packages/audioplayers_darwin/darwin/audioplayers_darwin/Sources/audioplayers_darwin/WrappedMediaPlayer.swift
@@ -1,5 +1,4 @@
 import AVKit
-import audioplayers_darwin_common
 
 private let defaultPlaybackRate: Double = 1.0
